hrono/5CQ/serviceone/requirements.txt
flask
redis
python-etcd
